pod waits for b
